requirements.txt
python-telegram-bot==20.6
supabase==2.4.0
Flask==3.0.2
qrcode
pillow